requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic[email]==2.7.0
pydantic-settings==2.2.1
sqlalchemy[asyncio]==2.0.29
aiosqlite==0.20.0
alembic==1.13.1
aioredis==2.0.1
kafka-python-ng==2.2.2
confluent-kafka==2.4.0
fastavro==1.9.4
mitmproxy==10.3.1
scapy==2.5.0
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
spacy==3.7.4
pyod==1.1.3
scikit-learn==1.4.2
prophet==1.1.5
pyyaml==6.0.1
httpx==0.27.0
jsonschema==4.22.0
detect-secrets==1.4.0
geoip2==4.8.0
structlog==24.1.0
prometheus-client==0.20.0
pytest==8.2.0
pytest-asyncio==0.23.6
slowapi==0.1.9
PyJWT==2.11.0
cryptography==46.0.4
python-jose==3.5.0
bcrypt==4.0.1
passlib==1.7.4
